sql: Refactor IdGen, add to StatementContext
Supporting CTEs requires either unification or translation of IDs
generated during planning vs. those generated during decorrelation.
This is due to the fact that references to CTEs use LocalIds,
the same which are generated during decorrelation. @benesch
:recommended unification, so that's what this PR does.

By hoisting IdGen into a StatementContext, we can use it to generate
local IDs for CTEs (done in subsequent commits), as well as pass
it along to the decorrelation process.

expr::IdGen gets refactored to an Rc'ed RefCell because being
honest with the borrow checker resulted in a very large diff that
required making changes to nearly every call site that accepts
StatementContext.