repr,sql: handle set operations on numerics another way
Rather than canonicalizing numeric data whenever we pack it into a Row,
instead only canonicalize numeric data before it heads into a set
operation, like UNION or DISTINCT. This means we preserve trailing zeros
whenever possible while still ensuring reasonable semantics for set
operations.

Note that we're still not quite PostgreSQL compatible on set operations.
When PostgreSQL encounters `DISTINCT ... VALUES (1.0), (1.00)`, it
promises to output `1.0` if you've specified an `ORDER BY`, because 1.0
appears first. But we'll output `1` because we always trim off trailing
zeros from all data that flows through a set operation.

Importantly, this approach should generalize reasonably well to the
`char` type, which has a similar issue around trailing whitespace.
`char` will be a bit harder, though, because we'll need to cast chars to
strings on the way in to set operations, and then cast them back to
chars on the way out.