pgwire: add test expressing update count semantics
Addresses comment from https://github.com/MaterializeInc/materialize/pull/7263#issuecomment-919780355

On an UPDATE operation, PostgreSQL's returns a count of the number
of rows input to the operation, which is the total number of rows
that underwent a change during the statement's execution. That is,
each row fed into the operator had its on-disk value ~changed,
even if the operation was a NOP.

On an UPDATE operation, Materialize returns a count of the number
of rows changed by the operation, from the perspective of the
number of memebers of the output set that changed. This, I believe,
is a more interesting number to users of Materialize and provides
a simple means of expressing the differing philosophies of PG and MZ.