coord: make catalog_transaction compute instance oblivious
Tweak the API of `catalog_transaction` so it is oblivious to compute
instances. Rather than providing the closure with a pre-formed
`DataflowBuilder`, it now provides the closure with a `CatalogTxn`
struct that can be used to construct a `DataflowBuilder` on any cluster,
should the closure so desire.

This makes it possible for one closure to conditionally construct a
`DataflowBuilder` or not, which is useful for commands like `CREATE
[MATERIALIZED] VIEW`, which may or may not need to create an index.