storage: close remap operator input when source input closes
For remap collections, we previously downgraded the CapabilitySet
when its input frontier was empty, but did not express that its
persist shard's write frontier closed. This has the unintended
consequence of preventing reads from the shard as its upper
remains at a fixed point in the past.

Instead, express that the shard's upper is closed and can be
read from at any time.