storage: close remap operator input when source input closes
For remap collections, we previously downgraded the CapabilitySet
when its input frontier was empty; this has the unintended
consequence of preventing the collection from being queried once
the input frontier is empty.

Instead, close the collection's write handle to prevent further
writes, but do not interfere with its ability to send data or have
its timestamp advanced.