storage: Progress collections track since of data collection
A progress collection's since can never reach the upper of the
data collection it represents; otherise, on restart, we would
not be able to perform remap operations.

To ensure this doesn't occur, advance the progress collection's
since in lock-step with the data collection's since (which can
also never reach the upper).