storage: simplify opening critical since handle
Previously, the convenience method for opening shard readers, writers
required that we move the handle into our epcoh and errored if the
handle was already in our epoch.

This change lets the operation succeed as long as the handles's
state upon return is in our epoch, irrespective of the handle's
original state.