sql: error if PgConfigOptionName::Details set on CREATE SOURCE
We previously filtered out previously set values of this option,
but we should instead ensure that it is not set. This reinforces
the notion that our initial snapshot of the PG schema on source
creation is the canonical one and e.g. its columns and ordering
are the one that we will honor in the face of schema changes.