pg sources: speculatively check PG catalog for missing DDL operations
The PG logical replication stream does not include dropped tables;
however, it does have a telltale sign of having elided something
important when it propagates an empty transaction, i.e.
```
BEGIN;COMMIT;
```

When we encounter this pattern, it's a good sign that we should
look at the upstream catalog for any changes.