pg sources: fix double-request during long txn sends
if the PG server receives multiple toki-postgres::standby_status_update
values while sending a large txn, it might send multiple instances
of that txn's data.

We had originally introduces the change that caused this issue
to simply establish a connection with the upstream PG database,
giving us a chance to break the connection if it seemed unhealthy.
We can instead just re-use the client and issue some other query
to the server that should not introduce side effects.