storage/pg: do not rewind definite errors
We currently emit a retraction for any definite errors that we
encounter in the "rewind" section of the WAL. However, we cannot
prove that the snapshot dataflow also produced an error for the
table. Imagine this history:

1. Snapshot begins.
2. Snapshot completes.
3. Replication dataflow begins.
4. Upstream table receives incompatible schema change.
5. Opening reoplication slot is delayed.
6. Replication begins, encounters incompatible schema change in
   "rewind" portion of WAL for incompatible table.

In this case, we would produce _and_ retract the definite error
if we rewound.

On the other hand, there is no downside to producing multiple
definite errors to the collection.

This also improves the ease of understanding the "ceasing" status,
which indicates that a collection will not receive further updates
because we can continue maintaining the invariant that the diff
of all data erors we produce is > 0.